# openapi.json
{
  "components": {
    "schemas": {
...
      "Candle": {
        "properties": {
          "close": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Close"
          },
          "date": {
            "title": "Date",
            "type": "string"
          },
          "high": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "High"
          },
          "low": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Low"
          },
          "open": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Open"
          },
          "volume": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Volume"
          }
        },
        "required": [
          "date",
          "open",
          "high",
          "low",
          "close",
          "volume"
        ],
        "title": "Candle",
        "type": "object"
      },
      "CandlesResponse": {
        "properties": {
          "candles": {
            "items": {
              "$ref": "#/components/schemas/Candle"
            },
            "title": "Candles",
            "type": "array"
          },
          "days": {
            "title": "Days",
            "type": "integer"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          }
        },
        "required": [
          "symbol",
          "days",
          "candles"
        ],
        "title": "CandlesResponse",
        "type": "object"
      },
...
      "DistributionPolicy": {
        "enum": [
          "ACCUMULATING",
          "DISTRIBUTING",
          "UNKNOWN"
        ],
        "title": "DistributionPolicy",
        "type": "string"
      },
...
      "FxEvidenceOut": {
        "description": "Shared public evidence shape for every base-currency analysis.",
        "properties": {
          "as_of": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "As Of"
          },
          "base_currency": {
            "title": "Base Currency",
            "type": "string"
          },
          "fetched_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fetched At"
          },
          "missing_currencies": {
            "items": {
              "type": "string"
            },
            "title": "Missing Currencies",
            "type": "array"
          },
          "note": {
            "title": "Note",
            "type": "string"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "status": {
            "enum": [
              "identity",
              "converted",
              "incomplete"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "base_currency",
          "source",
          "as_of",
          "fetched_at",
          "missing_currencies",
          "note"
        ],
        "title": "FxEvidenceOut",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "InstrumentResponse": {
        "properties": {
          "ann_vol": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ann Vol"
          },
          "as_of": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "As Of"
          },
          "beta": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Beta"
          },
          "beta_benchmark": {
            "title": "Beta Benchmark",
            "type": "string"
          },
          "con_id": {
            "title": "Con Id",
            "type": "integer"
          },
          "currency": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Currency"
          },
          "exchange": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exchange"
          },
          "high_52w": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "High 52W"
          },
          "industry": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Industry"
          },
          "isin": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Isin"
          },
          "issuer_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Issuer Id"
          },
          "last_close": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Close"
          },
          "local_symbol": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Local Symbol"
          },
          "long_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Long Name"
          },
          "low_52w": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Low 52W"
          },
          "pct_from_52w_high": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pct From 52W High"
          },
          "pct_from_52w_low": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pct From 52W Low"
          },
          "primary_exchange": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Primary Exchange"
          },
          "provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          },
          "region": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region"
          },
          "risk": {
            "$ref": "#/components/schemas/InstrumentRiskReadiness"
          },
          "risk_base_currency": {
            "title": "Risk Base Currency",
            "type": "string"
          },
          "risk_fx_as_of": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk Fx As Of"
          },
          "risk_fx_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk Fx Source"
          },
          "sec_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sec Type"
          },
          "stock_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stock Type"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          },
          "trading_class": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trading Class"
          },
          "ucits_profile": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UcitsEtfProfileV1"
              },
              {
                "type": "null"
              }
            ]
          },
          "ucits_profile_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ucits Profile Reason"
          },
          "ucits_profile_status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProfileFreshness"
              },
              {
                "type": "null"
              }
            ]
          },
          "valid_exchanges": {
            "items": {
              "type": "string"
            },
            "title": "Valid Exchanges",
            "type": "array"
          }
        },
        "required": [
          "symbol",
          "con_id",
          "long_name",
          "exchange",
          "currency",
          "sec_type",
          "industry",
          "region",
          "provider",
          "isin",
          "primary_exchange",
          "local_symbol",
          "trading_class",
          "stock_type",
          "valid_exchanges",
          "issuer_id",
          "ucits_profile_status",
          "ucits_profile_reason",
          "ucits_profile",
          "last_close",
          "high_52w",
          "low_52w",
          "pct_from_52w_high",
          "pct_from_52w_low",
          "ann_vol",
          "beta",
          "beta_benchmark",
          "risk_base_currency",
          "risk_fx_source",
          "risk_fx_as_of",
          "risk",
          "as_of"
        ],
        "title": "InstrumentResponse",
        "type": "object"
      },
      "InstrumentRiskReadiness": {
        "properties": {
          "base_currency": {
            "title": "Base Currency",
            "type": "string"
          },
          "benchmark": {
            "title": "Benchmark",
            "type": "string"
          },
          "fx": {
            "$ref": "#/components/schemas/FxEvidenceOut"
          },
          "note": {
            "title": "Note",
            "type": "string"
          },
          "reason": {
            "anyOf": [
              {
                "enum": [
                  "fx_unavailable",
                  "missing_benchmark",
                  "insufficient_history"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "status": {
            "enum": [
              "ready",
              "partial",
              "unavailable"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "reason",
          "benchmark",
          "base_currency",
          "fx",
          "note"
        ],
        "title": "InstrumentRiskReadiness",
        "type": "object"
      },
...
      "MetadataProvenanceV1": {
        "additionalProperties": false,
        "properties": {
          "fetched_at_utc": {
            "format": "date-time",
            "title": "Fetched At Utc",
            "type": "string"
          },
          "source": {
            "const": "justetf",
            "title": "Source",
            "type": "string"
          },
          "source_url": {
            "title": "Source Url",
            "type": "string"
          }
        },
        "required": [
          "source",
          "source_url",
          "fetched_at_utc"
        ],
        "title": "MetadataProvenanceV1",
        "type": "object"
      },
...
      "ProfileFreshness": {
        "enum": [
          "FRESH",
          "STALE",
          "MISSING"
        ],
        "title": "ProfileFreshness",
        "type": "string"
      },
...
      "UcitsEtfProfileV1": {
        "additionalProperties": false,
        "properties": {
          "benchmark_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Benchmark Name"
          },
          "distribution_policy": {
            "$ref": "#/components/schemas/DistributionPolicy"
          },
          "domicile": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Domicile"
          },
          "fund_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fund Name"
          },
          "isin": {
            "title": "Isin",
            "type": "string"
          },
          "issuer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Issuer"
          },
          "provenance": {
            "$ref": "#/components/schemas/MetadataProvenanceV1"
          },
          "replication_method": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Replication Method"
          },
          "schema_version": {
            "const": "ucits_etf_profile_v1",
            "title": "Schema Version",
            "type": "string"
          },
          "ter_pct": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ter Pct"
          }
        },
        "required": [
          "schema_version",
          "isin",
          "fund_name",
          "issuer",
          "domicile",
          "ter_pct",
          "distribution_policy",
          "replication_method",
          "benchmark_name",
          "provenance"
        ],
        "title": "UcitsEtfProfileV1",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "QuantMind API",
    "version": "0.4.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/instruments/{symbol}": {
      "get": {
        "operationId": "instrument_api_instruments__symbol__get",
        "parameters": [
          {
            "in": "path",
            "name": "symbol",
            "required": true,
            "schema": {
              "title": "Symbol",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstrumentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Instrument"
      }
    },
    "/api/instruments/{symbol}/candles": {
      "get": {
        "operationId": "candles_api_instruments__symbol__candles_get",
        "parameters": [
          {
            "in": "path",
            "name": "symbol",
            "required": true,
            "schema": {
              "title": "Symbol",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "days",
            "required": false,
            "schema": {
              "default": 180,
              "maximum": 3650,
              "minimum": 1,
              "title": "Days",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandlesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Candles"
      }
    },
...
  }
}